+++ tests/testthat/test-r-file-extension.R
@@ -1,6 +1,6 @@
 get_result <- function(res, check) res$result[res$check == check]
 
-test_that("r_file_extension fails when .r files exist", {
+test_that("r_file_extension fails when .r or .q files exist", {
   gp_res <- gp("bad_r_ext", checks = "r_file_extension")
   res <- results(gp_res)
   expect_false(get_result(res, "r_file_extension"))
@@ -8,6 +8,7 @@ test_that("r_file_extension fails when .r files exist", {
   pos <- failed_positions(gp_res)$r_file_extension
   filenames <- vapply(pos, `[[`, "", "filename")
   expect_true(any(grepl("bad_func\\.r$", filenames)))
+  expect_true(any(grepl("old_syntax\\.q$", filenames)))
 })
 
 test_that("r_file_extension passes when only .R files exist", {