+++ NEWS.md
@@ -5,7 +5,7 @@
 * goodpractice now uses cli, and no longer depends on crayon and clisymbols (@olivroy, #167).
 * If your editor supports it, goodpractice now prints clickable hyperlinks to console.
 - New `describe_check()` function to print descriptions of all implemented checks (@152)
-* New `r_file_extension` check: flags R scripts using lowercase `.r` instead of `.R` (#121).
+* New `r_file_extension` check: flags R scripts using `.r` or `.q` instead of `.R` (#121).
 
 # goodpractice 1.0.5